% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rml_g.R
\name{print.rml_g_fit}
\alias{print.rml_g_fit}
\title{Print an rml_g_fit object}
\usage{
\method{print}{rml_g_fit}(x, ...)
}
\arguments{
\item{x}{An \code{rml_g_fit} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an rml_g_fit object
}
