% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csrml.R
\name{csrml}
\alias{csrml}
\title{Cross-sectional Reconciliation with Machine Learning}
\usage{
csrml(
  hat,
  obs,
  base,
  agg_mat,
  features = "all",
  approach = "randomForest",
  params = NULL,
  tuning = NULL,
  fit = NULL,
  sntz = FALSE,
  seed = NULL
)
}
\arguments{
\item{hat}{todo}

\item{obs}{todo}

\item{base}{todo}

\item{agg_mat}{todo}

\item{features}{todo}

\item{approach}{todo}

\item{params}{todo}

\item{tuning}{todo (see [mlr3tuning::auto_tuner], except 'learner')}

\item{fit}{todo}

\item{sntz}{todo}

\item{seed}{todo}
}
\value{
Reconciled forecasts or Reconciled models
}
\description{
Cross-sectional Reconciliation with Machine Learning
}
\examples{
# (2 x 3) base forecasts matrix (simulated), Z = X + Y
base <- matrix(rnorm(6, mean = c(20, 10, 10)), 2, byrow = TRUE)
# (10 x 3) in-sample residuals matrix (simulated)
fitted <- matrix(rnorm(3*100, mean = c(20, 10, 10)), 100, byrow = TRUE)
obs <- matrix(rnorm(2*100, mean = c(10, 10)), 100, byrow = TRUE)
agg_mat <- t(c(1,1))

# randomForest (default)
## Option 1: one step
rml <- csrml(hat = fitted, obs = obs, base = base,
             agg_mat = agg_mat, seed = 1996) # Reconciled forecasts
model_fit <- recoinfo(rml, verbose = FALSE)$fit # Reconciled models

## Option 2: two steps
model_fit <- csrml(hat = fitted, obs = obs,
                   agg_mat = agg_mat, seed = 1996) # Reconciled models
rml <- csrml(base = base, agg_mat = agg_mat, fit = model_fit) # Reconciled forecasts

# randomForest with mlr3
## Option 1: one step
rml_mlr3 <- csrml(hat = fitted, obs = obs, base = base, agg_mat = agg_mat, seed = 1996,
                  approach = "mlr3") # Reconciled forecasts
model_fit_mlr3 <- recoinfo(rml, verbose = FALSE)$fit # Reconciled models

## Option 2: two steps
model_fit_mlr3 <- csrml(hat = fitted, obs = obs, agg_mat = agg_mat, seed = 1996,
                        approach = "mlr3") # Reconciled models

rml_refit <- csrml(base = base, agg_mat = agg_mat, fit = model_fit_mlr3) # Reconciled forecasts

}
