% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrml.R
\name{ctrml}
\alias{ctrml}
\title{Cross-temporal Reconciliation with Machine Learning}
\usage{
ctrml(
  hat,
  obs,
  base,
  agg_mat,
  agg_order,
  features = "all",
  approach = "randomForest",
  params = NULL,
  tuning = NULL,
  fit = NULL,
  tew = "sum",
  sntz = FALSE,
  seed = NULL
)
}
\arguments{
\item{hat}{todo}

\item{obs}{todo}

\item{base}{todo}

\item{agg_mat}{todo}

\item{agg_order}{todo}

\item{features}{todo}

\item{approach}{todo}

\item{params}{todo}

\item{tuning}{todo (see [mlr3tuning::auto_tuner], except 'learner')}

\item{fit}{todo}

\item{tew}{todo}

\item{sntz}{todo}

\item{seed}{todo}
}
\value{
todo
}
\description{
Cross-temporal Reconciliation with Machine Learning
}
\examples{
hat <- rnorm(7*10, rep(c(20, 10, 5), 10*c(1, 2, 4)))
obs <- rnorm(4*10, 5)
base <- rnorm(7*3, rep(c(20, 10, 5), 3*c(1, 2, 4)))

base <- rbind(base+100, base*10, base)
obs <- rbind(obs*10, obs)
hat <- rbind(hat+100, hat*10, hat)
m <- 4 # from quarterly to annual temporal aggregation
reco <- ctrml(hat = hat, obs = obs, base = base, agg_order = m, tew = "sum",
              agg_mat = t(c(1,1)), approach = "xgboost", seed = 123,
              #params = list(.key = "regr.xgboost"),
              features = "str")

reco <- ctrml(hat = hat, obs = obs, base = base, agg_order = m, tew = "sum",
              agg_mat = t(c(1,1)), approach = "lightgbm", seed = 123,
              #params = list(.key = "regr.xgboost"),
              features = "str")

# tweedie xgboost
reco <- ctrml(hat = hat, obs = obs, base = base, agg_order = m, tew = "sum",
              agg_mat = t(c(1,1)), approach = "xgboost", seed = 123,
              params =  list(
                eta = 0.3,
                colsample_bytree = 1,
                min_child_weight = 1,
                max_depth = 6,
                gamma = 0,
                subsample = 1,
                objective = "reg:tweedie",
                tweedie_variance_power = 1.5
              ),
              features = "str")

# Tuning mlr3
reco <- ctrml(hat = hat, obs = obs, base = base, agg_order = m, tew = "sum",
              agg_mat = t(c(1,1)), approach = "mlr3", seed = 123,
              params = list(.key = "regr.ranger",
                            mtry = paradox::to_tune(paradox::p_int(1, 4))),
              tuning = list(terminator = mlr3tuning::trm("evals", n_evals = 10)),
              features = "rtw-full")

}
