% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoRecoML-package.R
\docType{package}
\name{FoRecoML-package}
\alias{FoRecoML}
\alias{FoRecoML-package}
\title{FoRecoML: Forecast Reconciliation with Machine Learning}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danigiro/FoRecoML}
  \item \url{https://danigiro.github.io/FoRecoML}
  \item Report bugs at \url{https://github.com/danigiro/FoRecoML/issues}
}

}
\author{
\strong{Maintainer}: Daniele Girolimetto \email{daniele.girolimetto@unipd.it} (\href{https://orcid.org/0000-0001-9387-1232}{ORCID})

Authors:
\itemize{
  \item Yangzhuoran Fin Yang \email{yangyangzhuoran@gmail.com} (\href{https://orcid.org/0000-0002-1232-8017}{ORCID})
  \item Jeroen Rombouts \email{rombouts@essec.edu} (\href{https://orcid.org/0000-0003-2255-4875}{ORCID})
  \item Ines Wilms \email{i.wilms@maastrichtuniversity.nl} (\href{https://orcid.org/0000-0003-3269-4601}{ORCID})
}

}
\keyword{package}
