% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlReco-package.R
\docType{package}
\name{FoRecoML-package}
\alias{FoRecoML}
\alias{FoRecoML-package}
\alias{mlReco-package}
\title{FoRecoML: Forecast Reconciliation with Machine Learning}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{package}
