% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terml.R
\name{terml}
\alias{terml}
\title{Temporal Reconciliation with Machine Learning}
\usage{
terml(
  hat,
  obs,
  base,
  agg_order,
  features = "all",
  approach = "randomForest",
  params = NULL,
  tuning = NULL,
  fit = NULL,
  tew = "sum",
  sntz = FALSE,
  seed = NULL
)
}
\arguments{
\item{hat}{todo}

\item{obs}{todo}

\item{base}{todo}

\item{agg_order}{todo}

\item{features}{todo}

\item{approach}{todo}

\item{params}{todo}

\item{tuning}{todo (see [mlr3tuning::auto_tuner], except 'learner')}

\item{fit}{todo}

\item{tew}{todo}

\item{sntz}{todo}

\item{seed}{todo}
}
\value{
Reconciled forecasts or Reconciled models
}
\description{
Temporal Reconciliation with Machine Learning
}
\examples{
set.seed(123)
# (7 x 1) base forecasts vector (simulated), m = 4

hat <- rnorm(7*10, rep(c(20, 10, 5), 10*c(1, 2, 4)))
obs <- rnorm(4*10, 5)
base <- rnorm(7*2, rep(c(20, 10, 5), 2*c(1, 2, 4)))

m <- 4 # from quarterly to annual temporal aggregation
reco <- terml(hat = hat, obs = obs, base = base, agg_order = m,
              features = "all")

}
